requirements.txt
app==0.0.1
beautifulsoup4==4.13.4
curl_cffi==0.10.0
Flask==3.1.1
joblib==1.4.2
numpy==2.3.1
pandas==2.3.0
psycopg2==2.9.10
python-dotenv==1.1.1
python_dateutil==2.9.0.post0
Requests==2.32.4
